Move update_user to handle it more seamlessly
