+++ framework/src/context.rs
@@ -1,26 +1,21 @@
 use chacha20poly1305::ChaCha20Poly1305;
 use dashmap::{DashMap, DashSet};
-use itertools::Itertools;
 use patreon::Client as Patreon;
 use roblox::Client as Roblox;
 use rowifi_cache::{Cache, CachedGuild, CachedMember};
 use rowifi_database::Database;
 use rowifi_models::{
-    bind::Bind,
     discord::{
         channel::embed::Embed,
         id::{ChannelId, GuildId, InteractionId, MessageId, RoleId, UserId, WebhookId},
         user::User,
     },
-    guild::{BlacklistActionType, RoGuild},
-    roblox::id::{AssetId as RobloxAssetId, UserId as RobloxUserId},
-    rolang::RoCommandUser,
     stats::BotStats,
     user::RoGuildUser,
 };
 use std::{
     borrow::ToOwned,
-    collections::{HashMap, HashSet},
+    collections::HashMap,
     ops::Deref,
     sync::{atomic::AtomicBool, Arc},
 };
@@ -30,7 +25,7 @@ use twilight_standby::Standby;
 use twilight_util::link::webhook;
 
 use crate::{
-    error::{CommandError, RoError},
+    error::RoError,
     respond::Responder,
 };
 
@@ -246,197 +241,6 @@ impl CommandContext {
 }
 
 impl BotContext {
-    #[allow(clippy::needless_collect)]
-    pub async fn update_user(
-        &self,
-        member: Arc<CachedMember>,
-        user: &RoGuildUser,
-        server: &CachedGuild,
-        guild: &RoGuild,
-        guild_roles: &HashSet<RoleId>,
-        bypass_roblox_cache: bool,
-    ) -> Result<(Vec<RoleId>, Vec<RoleId>, String), RoError> {
-        let mut added_roles = Vec::<RoleId>::new();
-        let mut removed_roles = Vec::<RoleId>::new();
-
-        if let Some(verification_role) = guild.verification_role {
-            let verification_role = RoleId::new(verification_role as u64).unwrap();
-            if guild_roles.get(&verification_role).is_some()
-                && member.roles.contains(&verification_role)
-            {
-                removed_roles.push(verification_role);
-            }
-        }
-
-        if let Some(verified_role) = guild.verified_role {
-            let verified_role = RoleId::new(verified_role as u64).unwrap();
-            if guild_roles.get(&verified_role).is_some() && !member.roles.contains(&verified_role) {
-                added_roles.push(verified_role);
-            }
-        }
-
-        let user_id = RobloxUserId(user.roblox_id as u64);
-        let user_roles = self
-            .roblox
-            .get_user_roles(user_id)
-            .await?
-            .iter()
-            .map(|r| (r.group.id.0 as i64, r.role.rank as i64))
-            .collect::<HashMap<_, _>>();
-        let roblox_user = self.roblox.get_user(user_id, bypass_roblox_cache).await?;
-        let command_user = RoCommandUser {
-            user,
-            roles: &member.roles,
-            ranks: &user_roles,
-            username: &roblox_user.name,
-        };
-
-        if !guild.blacklists.is_empty() {
-            let success = guild
-                .blacklists
-                .iter()
-                .find(|b| b.evaluate(&command_user).unwrap());
-            if let Some(success) = success {
-                match guild.settings.blacklist_action {
-                    BlacklistActionType::None => {}
-                    BlacklistActionType::Kick => {
-                        let _ = self
-                            .http
-                            .remove_guild_member(server.id, member.user.id)
-                            .exec()
-                            .await;
-                    }
-                    BlacklistActionType::Ban => {
-                        let _ = self.http.create_ban(server.id, member.user.id).exec().await;
-                    }
-                };
-                return Err(CommandError::Other(success.reason.clone()).into());
-            }
-        }
-
-        let mut nick_bind: Option<&dyn Bind> = None;
-        let mut roles_to_add = Vec::new();
-
-        for r in &guild.rankbinds {
-            let to_add = match user_roles.get(&r.group_id) {
-                Some(rank_id) => *rank_id == r.rank_id as i64,
-                None => r.rank_id == 0,
-            };
-            if to_add {
-                if let Some(highest) = nick_bind {
-                    if highest.priority() < r.priority() {
-                        nick_bind = Some(r);
-                    }
-                } else {
-                    nick_bind = Some(r);
-                }
-                roles_to_add.extend(r.discord_roles.iter().copied());
-            }
-        }
-
-        for g in &guild.groupbinds {
-            if user_roles.contains_key(&g.group_id) {
-                if let Some(highest) = nick_bind {
-                    if highest.priority() < g.priority() {
-                        nick_bind = Some(g);
-                    }
-                } else {
-                    nick_bind = Some(g);
-                }
-                roles_to_add.extend(g.discord_roles.iter().copied());
-            }
-        }
-
-        for c in &guild.custombinds {
-            if c.command.evaluate(&command_user).unwrap() {
-                if let Some(highest) = nick_bind {
-                    if highest.priority() < c.priority() {
-                        nick_bind = Some(c);
-                    }
-                } else {
-                    nick_bind = Some(c);
-                }
-                roles_to_add.extend(c.discord_roles.iter().copied());
-            }
-        }
-
-        for a in &guild.assetbinds {
-            if self
-                .roblox
-                .get_asset(
-                    user_id,
-                    RobloxAssetId(a.id as u64),
-                    &a.asset_type.to_string(),
-                )
-                .await?
-                .is_some()
-            {
-                if let Some(highest) = nick_bind {
-                    if highest.priority() < a.priority() {
-                        nick_bind = Some(a);
-                    }
-                } else {
-                    nick_bind = Some(a);
-                }
-                roles_to_add.extend(a.discord_roles.iter().copied());
-            }
-        }
-
-        for bind_role in &guild.all_roles {
-            let r = RoleId::new(*bind_role as u64).unwrap();
-            if guild_roles.get(&r).is_some() {
-                if roles_to_add.contains(bind_role) {
-                    if !member.roles.contains(&r) {
-                        added_roles.push(r);
-                    }
-                } else if member.roles.contains(&r) {
-                    removed_roles.push(r);
-                }
-            }
-        }
-
-        let original_nick = member
-            .nick
-            .as_ref()
-            .map_or_else(|| member.user.name.as_str(), |s| s.as_str());
-        let nick_bypass = self.has_nickname_bypass(server, &member);
-        let nickname = if nick_bypass {
-            original_nick.to_string()
-        } else {
-            nick_bind.map_or_else(
-                || roblox_user.name.to_string(),
-                |nick_bind| nick_bind.nickname(&roblox_user, user, &member.user.name, &member.nick),
-            )
-        };
-
-        if nickname.len() > 32 {
-            return Err(CommandError::Other(format!(
-                "The supposed nickname {} was found to be more than 32 characters",
-                nickname
-            )).into());
-        }
-
-        let update = self.http.update_guild_member(server.id, member.user.id);
-        let role_changes = !added_roles.is_empty() || !removed_roles.is_empty();
-        let mut roles = member.roles.clone();
-        roles.extend_from_slice(&added_roles);
-        roles.retain(|r| !removed_roles.contains(r));
-        roles = roles.into_iter().unique().collect::<Vec<RoleId>>();
-
-        let nick_changes = nickname != original_nick;
-
-        if role_changes || nick_changes {
-            update
-                .roles(&roles)
-                .nick(Some(&nickname))
-                .unwrap()
-                .exec()
-                .await?;
-        }
-
-        Ok((added_roles, removed_roles, nickname))
-    }
-
     #[allow(clippy::cast_possible_wrap)]
     pub async fn get_linked_user(
         &self,
