+++ rowifi/src/utils/mod.rs
@@ -1,0 +1,3 @@
+mod update_user;
+
+pub use update_user::*;
